+++ packages/terra-framework-docs/CHANGELOG.md
@@ -14,7 +14,7 @@
   * Migrate docs from terra-navigation-layout
   * Migrate docs from terra-menu
   * Migrate docs from terra-notification-dialog.
-
+  * Migrate docs from terra-hookshot
 
 ## 1.7.0 - (March 8, 2022)